refactor(ses): Detangle shims from modules
