+++ packages/ses/src/module-shim.js
@@ -13,8 +13,6 @@ import {
 import { load } from './module-load.js';
 import { link } from './module-link.js';
 import { getDeferredExports } from './module-proxy.js';
-import { getGlobalIntrinsics } from './intrinsics.js';
-import { tameFunctionToString } from './tame-function-tostring.js';
 import { InertCompartment, InertStaticModuleRecord } from './inert.js';
 import {
   CompartmentPrototype,
@@ -167,6 +165,8 @@ defineProperties(
   getOwnPropertyDescriptors(ModularCompartmentPrototypeExtension),
 );
 
+export { CompartmentPrototype };
+
 export const makeModularCompartmentConstructor = (
   targetMakeCompartmentConstructor,
   intrinsics,
@@ -247,20 +247,4 @@ export const makeModularCompartmentConstructor = (
   ModularCompartment.prototype = CompartmentPrototype;
 
   return ModularCompartment;
-};
-
-// TODO wasteful to do it twice, once before lockdown and again during
-// lockdown. The second is doubly indirect. We should at least flatten that.
-const nativeBrander = tameFunctionToString();
-
-export const ModularCompartment = makeModularCompartmentConstructor(
-  makeModularCompartmentConstructor,
-  getGlobalIntrinsics(globalThis),
-  nativeBrander,
-);
-
-export {
-  ModularCompartment as Compartment,
-  CompartmentPrototype,
-  makeModularCompartmentConstructor as makeCompartmentConstructor,
 };
