+++ packages/ses/test/global-object-properties.test.js
@@ -1,6 +1,6 @@
 import tap from 'tap';
 import sinon from 'sinon';
-import { Compartment } from '../src/compartment-shim.js';
+import '../lockdown.js';
 import stubFunctionConstructors from './stub-function-constructors.js';
 
 const { test } = tap;
