+++ packages/ses/test/import-gauntlet.test.js
@@ -2,7 +2,7 @@
 // modules using a single Compartment.
 
 import tap from 'tap';
-import { Compartment } from '../src/module-shim.js';
+import '../ses.js';
 import { resolveNode, makeNodeImporter } from './node.js';
 
 const { test } = tap;
